+++ cmake/nfxGraphicsSources.cmake
@@ -3,6 +3,7 @@
 #==============================================================================
 
 set(NFX_GRAPHICS_SOURCES
+    src/gl/material/Material.cpp
     src/gl/mesh/Mesh.cpp
     src/gl/mesh/Primitive.cpp
     src/gl/pipeline/RenderState.cpp